+++ apps/web/components/index.ts
@@ -1,0 +1,5 @@
+// UI components
+export { Testimonials } from './ui/Testimonials';
+
+// Re-export types
+export type { Testimonial } from '../data/testimonials';
